Cleanup new test case
