+++ burn-fusion/src/stream/execution/explorer.rs
