+++ burn-fusion/src/stream/execution/mod.rs
@@ -1,9 +1,12 @@
 mod base;
-mod exploration;
+mod explorer;
 mod policy;
 mod processor;
 
 pub(crate) use base::*;
-pub(crate) use exploration::*;
+pub(crate) use explorer::*;
 pub(crate) use policy::*;
 pub(crate) use processor::*;
+
+#[cfg(test)]
+mod tests;
