+++ burn-fusion/src/stream/multi.rs
@@ -1,5 +1,5 @@
 use super::{
-    execution::{ExecutionMode, Process, Processor},
+    execution::{ExecutionMode, Processor, StreamSegment},
     store::{ExplorationId, ExplorationStore},
     Operation, OperationDescription, OperationQueue,
 };
@@ -11,27 +11,6 @@ use crate::{FusionBackend, HandleContainer};
 pub struct MultiStream<B: FusionBackend> {
     streams: Vec<Stream<B>>,
     optimizations: ExplorationStore<B::Optimization>,
-}
-
-struct Stream<B: FusionBackend> {
-    queue: OperationQueue<B>,
-    processor: Processor<B::Optimization>,
-}
-
-#[derive(new)]
-struct StreamExecutionProcess<'a, B: FusionBackend> {
-    queue: &'a mut OperationQueue<B>,
-    handles: &'a mut HandleContainer<B>,
-}
-
-impl<'i, B: FusionBackend> Process<B::Optimization> for StreamExecutionProcess<'i, B> {
-    fn operations<'a>(&'a self) -> &[OperationDescription] {
-        &self.queue.relative
-    }
-
-    fn execute(&mut self, id: ExplorationId, store: &mut ExplorationStore<B::Optimization>) {
-        self.queue.execute(id, self.handles, store)
-    }
 }
 
 impl<B: FusionBackend> MultiStream<B> {
@@ -53,7 +32,7 @@ impl<B: FusionBackend> MultiStream<B> {
         if let Some(item) = self.streams.first_mut() {
             item.queue.add(desc, operation);
             item.processor.process(
-                StreamExecutionProcess::new(&mut item.queue, handles),
+                Segment::new(&mut item.queue, handles),
                 &mut self.optimizations,
                 ExecutionMode::Lazy,
             );
@@ -64,11 +43,32 @@ impl<B: FusionBackend> MultiStream<B> {
     pub fn drain(&mut self, handles: &mut HandleContainer<B>) {
         self.streams.iter_mut().for_each(|item| {
             item.processor.process(
-                StreamExecutionProcess::new(&mut item.queue, handles),
+                Segment::new(&mut item.queue, handles),
                 &mut self.optimizations,
                 ExecutionMode::Sync,
             );
         });
+    }
+}
+
+struct Stream<B: FusionBackend> {
+    queue: OperationQueue<B>,
+    processor: Processor<B::Optimization>,
+}
+
+#[derive(new)]
+struct Segment<'a, B: FusionBackend> {
+    queue: &'a mut OperationQueue<B>,
+    handles: &'a mut HandleContainer<B>,
+}
+
+impl<'i, B: FusionBackend> StreamSegment<B::Optimization> for Segment<'i, B> {
+    fn operations<'a>(&'a self) -> &[OperationDescription] {
+        &self.queue.relative
+    }
+
+    fn execute(&mut self, id: ExplorationId, store: &mut ExplorationStore<B::Optimization>) {
+        self.queue.execute(id, self.handles, store)
     }
 }
